services/cfx-router/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pyyaml>=6.0.1
httpx>=0.25.0
supabase>=2.0.0

